deduplicate DeviceBufferManager copypaste - apply_avail_min is all that separated them
