+++ src/alsa_backend/device.rs
@@ -43,9 +43,8 @@ use crate::PrcFmt;
 use crate::ProcessingState;
 use crate::Res;
 use crate::StatusMessage;
-use crate::alsa_backend::buffermanager::{
-    CaptureBufferManager, DeviceBufferManager, PlaybackBufferManager,
-};
+use crate::alsa_backend::buffermanager::{CaptureBufferManager, PlaybackBufferManager};
+use crate::alsa_backend::device_buffer_manager::DeviceBufferManager;
 use crate::alsa_backend::utils::{
     CaptureElements, CaptureParams, CaptureResult, ElemData, FileDescriptors, PlaybackParams,
     find_elem, process_events, recover_suspended_pcm, state_desc, sync_linked_controls,
