+++ src/signals/mod.rs
@@ -3,7 +3,7 @@ mod unix_signals;
 #[cfg(windows)]
 mod windows_signals;
 
-pub(self) const EXIT_FORCED: i32 = 103; // Exit was forced by a second SIGINT
+const EXIT_FORCED: i32 = 103; // Exit was forced by a second SIGINT
 
 #[cfg(not(windows))]
 pub use unix_signals::handle_signals;
